chore: dir rename
