+++ packages/edition-slices/src/configured-tiles/tile-vertical-a/index.tsx
