+++ packages/edition-slices/src/configured-tiles/tile-vertical-a/styles.tsx
