+++ packages/edition-slices/src/slices/leadoneandone/index.js
@@ -1,9 +1,9 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { LeadOneAndOneSlice } from "@times-components-native/slice-layout";
-import { TileA, TileB, TileU, TileZ, TileAF } from "../../tiles";
+import { TileA, TileB, TileU, TileZ } from "../../tiles";
 import { ResponsiveSlice } from "../shared";
-import TileVerticalA from "@times-components-native/edition-slices/src/flexible-tiles/tile-vertical-a";
+import TileVerticalA from "@times-components-native/edition-slices/src/configured-tiles/tile-vertical-a";
 
 class LeadOneAndOne extends Component {
   constructor(props) {
